Crate ilockmvp

INTERLOCK NETWORK MVP SMART CONTRACT

PSP22 TOKEN

REWARDS

This is a standard ERC20-style token contract with provisions for enforcing a token distribution vesting schedule, and for rewarding interlockers for browsing the internet with the Interlock browser extension.

To ensure build with cargo-contract version 2.0.0, run:

cargo install cargo-contract –force –version 2.0.0

To build, run:

cargo +nightly contract build

To build docs, run:

cargo +nightly doc –no-deps –document-private-items –open

To reroute docs in Github, run:

echo “<meta http-equiv="refresh" content="0; url=ilockmvp">” > target/doc/index.html; cp -r target/doc ./docs

Re-exports

pub use self::ilockmvp::ILOCKmvp;

pub use self::ilockmvp::ILOCKmvpRef;

Modules

ilockmvp