Module ilockmvp::ilockmvp

Structs

AccountID

This is a type wrapper to implement Default method on AccountId type. Ink 4 stable eliminated AccountId Default (which was zero address, that has known private key) …we only really need this because Openbrush contract relies on deriving Default for contract storage, and our AccesData struct contains AccountId.

AppData

Approval

Specify approval event.

CallBuilder

The ink! smart contract’s call builder.

Check 🔒

ILOCKmvp

ILOCKmvp struct contains overall storage data for contract

ILOCKmvpRef

ILOCKmvp struct contains overall storage data for contract

MultisigData

Port

Information pertaining to port definition in application/socket/port contract connectivity formalism.

Reward

Specify reward event.

RewardData

Signature

TransactionData struct contains all pertinent information for multisigtx transaction

Socket

Information pertaining to socket definition in application/socket/port contract connectivity formalism.

Transaction

TransactionData struct contains all pertinent information for multisigtx transaction

Transfer

Specify transfer event.

Enums

OtherError

Other contract error types.

__ink_ConstructorDecoder

__ink_EventBase

__ink_MessageDecoder

__ink_UndefinedAmountOfTopics

Constants

ADD_SIGNATORY

APP_KEY

This is upgradable storage for the application connection feature of this PSP22 contract (ie, the application/socket/port contract connectivity formalism).

ARR_LEN 🔒

ARR_LEN 🔒

ARR_LEN 🔒

CHANGE_THRESHOLD

CHANGE_TIMELIMIT

CONCATP_NHPMWYD3NJA 🔒

CONCATP_NHPMWYD3NJA 🔒

CONCATP_NHPMWYD3NJA 🔒

CONCAT_ARR 🔒

CONCAT_ARR 🔒

CONCAT_ARR 🔒

CONCAT_STR 🔒

CONCAT_STR 🔒

CONCAT_STR 🔒

CONSTRUCTOR_0 🔒

CREATE_PORT

DECIMALS_POWER10

ID_LENGTH

Magic numbers.

INITIAL_REWARDS

MESSAGE_0 🔒

MESSAGE_1 🔒

MESSAGE_2 🔒

MESSAGE_3 🔒

MESSAGE_4 🔒

MESSAGE_5 🔒

MESSAGE_6 🔒

MESSAGE_7 🔒

MESSAGE_8 🔒

MESSAGE_9 🔒

MESSAGE_10 🔒

MESSAGE_11 🔒

MESSAGE_12 🔒

MESSAGE_13 🔒

MESSAGE_14 🔒

MESSAGE_15 🔒

MESSAGE_16 🔒

MESSAGE_17 🔒

MESSAGE_18 🔒

MESSAGE_19 🔒

MESSAGE_20 🔒

MESSAGE_21 🔒

MESSAGE_22 🔒

MESSAGE_23 🔒

MESSAGE_24 🔒

MESSAGE_25 🔒

MESSAGE_26 🔒

MESSAGE_27 🔒

MESSAGE_28 🔒

MESSAGE_29 🔒

MESSAGE_30 🔒

MESSAGE_31 🔒

MESSAGE_32 🔒

MESSAGE_33 🔒

MESSAGE_34 🔒

MESSAGE_35 🔒

MESSAGE_36 🔒

MESSAGE_37 🔒

MESSAGE_38 🔒

MIN_SHARE

MULTISIG_KEY

This is upgradable storage for the multisig feature of this PSP22 contract (ie, the application/socket/port contract connectivity formalism).

MULTISIG_TIME

ONE_MONTH

POOL_COUNT

REMOVE_SIGNATORY

REWARD_KEY

This is upgradable storage for the token rewarding feature of this PSP22 contract.

SUPPLY_CAP

THRESHOLD_MIN

TIME_LIMIT_MIN

TOKEN_CAP

Token data.

TOKEN_DECIMALS

TOKEN_NAME

TOKEN_SYMBOL

TRANSFER_OWNERSHIP

Multisig functions.

UNPAUSE

UPDATE_CONTRACT

Functions

__ink_generate_metadata

call 🔒

deploy 🔒

push_contract 🔒

Type Definitions

AccountId 🔒

Balance 🔒

BlockNumber 🔒

Environment 🔒

Event

Needed for Openbrush internal event emission implementations.

Hash 🔒

OtherResult

Convenience Result Type

PSP22Result

Convenience Result Type.

Timestamp 🔒