Enum ilockmvp::ilockmvp::OtherError

pub enum OtherError {

    CallerNotOwner,
    StakeholderSharePaid,
    StakeholderNotFound,
    CliffNotPassed,
    PayoutTooEarly,
    PaymentTooLarge,
    NoSocket,
    NoPort,
    NotContract,
    PortLocked,
    PortCapSurpassed,
    CannotRewardContract,
    UnsafeContract,
    CallerNotOperator,
    Overflow,
    Underflow,
    DivError,
    ShareTooSmall,
    InvalidPool,
    InvalidPort,
    Custom(String),

}

Other contract error types.

Variants

CallerNotOwner

Returned if caller is not contract owner.

StakeholderSharePaid

Returned if stakeholder share is entirely paid out.

StakeholderNotFound

Returned if the stakeholder doesn’t exist.

CliffNotPassed

Returned if stakeholder has not yet passed cliff.

PayoutTooEarly

Returned if it is too soon to payout for month.

PaymentTooLarge

Returned if reward is too large.

NoSocket

Returned if socket does not exist.

NoPort

Returned if port does not exist.

NotContract

Returned if not contract.

PortLocked

Returned if only owner can add socket.

PortCapSurpassed

Returned if port cap is surpassed.

CannotRewardContract

Returned if reward recipient is a contract.

UnsafeContract

Returned if socket contract does not match registered hash.

CallerNotOperator

Returned if application contract caller is not its operator.

Overflow

Returned if checked add overflows.

Underflow

Returned if checked sub underflows.

DivError

Returned if checked divide errors out.

ShareTooSmall

Returned if share is not greater than zero.

InvalidPool

Returned if pool number provided is invalid.

InvalidPort

Returned if port number provided is invalid.

Custom(String)

Custom contract error.