Struct ilockmvp::ilockmvp::AppData

pub struct AppData {
    pub ports: <Mapping<u16, Port> as AutoStorableHint<ManualKey<{ _ }, ManualKey<APP_KEY>>>>::Type,
    pub sockets: <Mapping<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, Socket> as AutoStorableHint<ManualKey<{ _ }, ManualKey<APP_KEY>>>>::Type,
    pub _reserved: <Option<()> as AutoStorableHint<ManualKey<{ _ }, ManualKey<APP_KEY>>>>::Type,
}

Fields
ports: <Mapping<u16, Port> as AutoStorableHint<ManualKey<{ _ }, ManualKey<APP_KEY>>>>::Type

Contains information specifying a particular type of connecting external application contract via the application/socket abstraction.

When an application contract creates a connecting socket with this token contract with a particular port, it adheres to the logic and protocol specified by the port type.

For example, PORT 0 in this contract only accepts connections from universal access NFT contract owned by Interlock, and for every socket call from a UANFT contract application, tokens in the amount of the set NFT price are transferred from the calling minter to this ILOCK contractโs owner account. On the application side, once the ILOCK tokens are successfully transferred via the port protocol, a UANFT is minted to the caller.

For example, PORT 1 in this contract is the same as PORT 0, but UANFT application contracts are owned by different operators, and on each socket call, the protocol includes an additional tax in ILOCK, which Interlock Network collects.

The mapping is from port number, to port details and specs.

Only this contractโs owner has the authority to create or edit a port.

See detailed struct below.

ports: port number -> port(app contract hash, metadata, port owner)

sockets: <Mapping<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, Socket> as AutoStorableHint<ManualKey<{ _ }, ManualKey<APP_KEY>>>>::Type

Contains information specifying a particular instance of an application (as defined by port application hash) contractโs connection to this PSP22 contract.

Similar to the standard TCP/IP address:port format, the port specifies the protocol, and the address specifies the operator of that particular instance of the application contract connecting to this PSP22 contract.

In the example of PORT 1, the address of a socket connection is the address that receives the ILOCK token transfer, ultimately in exchange for the UANFT mint back on the application side.

The mapping is from application address, to socket operator address and port number.

One socket may serve multiple applications (ie, the same operator address:port number pair) which is a slight deviation from the socket formality in TCP/IP.

Any agent with a verified application contract may connect to this PSP22 contract without permission from this contractโs owner.

See detailed struct below.

sockets: application contract address -> socket(app operator address : port)

_reserved: <Option<()> as AutoStorableHint<ManualKey<{ _ }, ManualKey<APP_KEY>>>>::Type

Expand storage related to the application/socket/port functionality.