Struct ilockmvp::ilockmvp::CallBuilder

#[repr(transparent)]
pub struct CallBuilder {
    account_id: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId,
}

The ink! smart contract’s call builder.

Implements the underlying on-chain calling of the ink! smart contract messages and trait implementations in a type safe way.

Fields
account_id: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId
Implementations

impl CallBuilder

pub fn check_time( &mut self ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<PSP22Result<()>>>>

Function to check if enough time has passed to collect next payout.

This function ensures Interlock cannot rush the vesting schedule.

This function must be called before the next round of token distributions.

pub fn remaining_time( &self ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<<<ILOCKmvp as ContractEnv>::Env as Environment>::Timestamp>>>

Time in seconds until next payout in minutes.

pub fn register_stakeholder( &mut self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, __ink_binding_1: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, __ink_binding_2: u8 ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<u8>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>>>, Set<ReturnType<PSP22Result<()>>>>

Function that registers a stakeholder’s wallet and vesting info.

Data used to calculate monthly payouts and track net paid.

Stakeholder data also used for stakeholder to verify their place in vesting schedule.

pub fn stakeholder_data( &self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>, Set<ReturnType<(StakeholderData, <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, String)>>>

Function that returns a stakeholder’s payout and other data.

This will allow stakeholders to verify their stake from explorer if so motivated.

Returns tuple (StakeholderData, payremaining, payamount, poolname).

pub fn distribute_tokens( &mut self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>, Set<ReturnType<OtherResult<()>>>>

General function to transfer share a stakeholder is currently entitled to.

This is called once per stakeholder per month by Interlock, Interlock paying fees.

Pools are guaranteed to have enough tokens for all stakeholders.

pub fn payout_tokens( &mut self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, __ink_binding_1: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, __ink_binding_2: String ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>>>, Set<ReturnType<OtherResult<()>>>>

Function used to payout tokens to pools with no vesting schedule. POOL ARGUMENTS: PARTNERS COMMUNITY PUBLIC PROCEEDS

pub fn pool_data( &self, __ink_binding_0: u8 ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<u8>, EmptyArgumentList>>>, Set<ReturnType<(String, String, String, String)>>>

Function that returns pool data in human readable format..

This will allow observers to verify vesting parameters for each pool (esp. theirs).

Observers may verify pool data from explorer if so motivated.

Pool numbers range from 0-11.

Returns (name, tokens, vests, cliff) (formatted for convenient for Substrate UI)..

pub fn pool_balance( &self, __ink_binding_0: u8 ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<u8>, EmptyArgumentList>>>, Set<ReturnType<(String, <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance)>>>

Get current balance of any vesting pool.

Provide human readable and numberic format.

pub fn reward_interlocker( &mut self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, __ink_binding_1: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>, EmptyArgumentList>>>>, Set<ReturnType<OtherResult<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>>>>

Reward the interlocker for browsing, etc.

This is a manual rewarding function, to override the socket formalism.

pub fn rewarded_interlocker_total( &self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>, Set<ReturnType<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>>>

Get amount rewarded to interlocker to date.

pub fn rewarded_total( &self ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>>>

Get total amount rewarded to date.

pub fn months_passed( &self ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<u16>>>

Function to get the number of months passed for contract.

pub fn cap( &self ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>>>

Function to get the supply cap minted on TGE.

pub fn update_contract( &mut self, __ink_binding_0: [u8; 32] ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<[u8; 32]>, EmptyArgumentList>>>, Set<ReturnType<PSP22Result<()>>>>

Modifies the code which is used to execute calls to this contract address.

This upgrades the token contract logic while using old state.

pub fn create_port( &mut self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::Hash, __ink_binding_1: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, __ink_binding_2: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, __ink_binding_3: bool, __ink_binding_4: u16, __ink_binding_5: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, ArgumentList<Argument<u16>, ArgumentList<Argument<bool>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::Hash>, EmptyArgumentList>>>>>>>>, Set<ReturnType<PSP22Result<()>>>>

Create a new port that application contract can register with.

Each port tracks amount rewarded, tax collected, if it is locked or not, owner.

A locked port may only be registered by the Interlock Network foundation.

pub fn create_socket( &mut self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, __ink_binding_1: u16 ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<u16>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>>, Set<ReturnType<OtherResult<()>>>>

Rewards/staking/application contracts register with this token contract here.

Contract must first register with token contract as port to allow connection via socket (ie, a port must first exist before a socket may form)..

pub fn call_socket( &mut self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, __ink_binding_1: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, __ink_binding_2: Vec<u8> ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<Vec<u8>>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>>>, Set<ReturnType<OtherResult<()>>>>

Check for socket and apply custom logic after being called from application contract.

Application contract calls its socket to trigger internal logic defined by port.

Default parameters are address and amount or value.

Additional parameters may be encoded as _data: Vec.

pub fn socket( &self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>, Set<ReturnType<Socket>>>

Get socket info.

pub fn port( &self, __ink_binding_0: u16 ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<u16>, EmptyArgumentList>>>, Set<ReturnType<Port>>>

Get port info.

pub fn TESTING_increment_month( &mut self ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<OtherResult<bool>>>>

Function to increment monthspassed for testing.

MUST BE DELETED PRIOR TO TGE