Struct ilockmvp::ilockmvp::ILOCKmvp

pub struct ILOCKmvp {
    pub psp22: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    pub ownable: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    pub metadata: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    pub pausable: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    pub reward: <RewardData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    pub app: <AppData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    pub multisig: <MultisigData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    pub rewards: <<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    pub circulating: <<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    pub proceeds: <<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
}

ILOCKmvp struct contains overall storage data for contract

Fields
psp22: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

Openbrush PSP22.

ownable: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

Openbrush ownership extension.

metadata: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

Openbrush metadata extension.

pausable: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

Openbrush pausable extensios.

reward: <RewardData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

ILOCK Rewards info.

app: <AppData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

ILOCK connecting application contract info.

multisig: <MultisigData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

ILOCK multisig contract info.

rewards: <<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

ILOCK token pool balances.

circulating: <<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance as AutoStorableHint<ManualKey<{ _ }, ()>>>::Typeproceeds: <<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type
Implementations

impl ILOCKmvp

pub fn new_token( timelimit: <<ILOCKmvp as ContractEnv>::Env as Environment>::Timestamp, rewardmax: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, signatory_2: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, signatory_3: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> OtherResult<Self>

Constructor to initialize contract.

pub fn order_multisigtx(&mut self, function: String) -> OtherResult<()>

Function to order multisigtx transaction.

pub fn sign_multisigtx(&mut self, function: String) -> OtherResult<()>

A multisigtx signer calls this to sign.

pub fn add_signatory( &mut self, signatory: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, function: String ) -> OtherResult<()>

This adds a signatory from the list of permitted signatories.

pub fn remove_signatory( &mut self, signatory: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, function: String ) -> OtherResult<()>

This removes a signatory from the list of permitted signatories.

pub fn change_threshold( &mut self, threshold: u16, function: String ) -> OtherResult<()>

This changes signer threshold for approving multisigtx.

pub fn change_multisigtxtimelimit( &mut self, timelimit: <<ILOCKmvp as ContractEnv>::Env as Environment>::Timestamp, function: String ) -> OtherResult<()>

This modifies timelimit for a multisig transaction.

pub fn threshold(&self) -> u16

This gets the current signature threshold for multisigtx.

pub fn multisigtimelimit( &self ) -> <<ILOCKmvp as ContractEnv>::Env as Environment>::Timestamp

This gets the current timelimit for signatories to sign multisigtx.

pub fn signatories(&mut self) -> OtherResult<Vec<AccountID>>

This gets a list of current accounts permitted to sign multisigtx.

pub fn signatory_count(&self) -> u8

This gets number of signatories permitted to sign multisigtx.

pub fn signature_count(&self) -> u8

This gets current number of signatures for multisigtx.

pub fn check_signatures(&mut self) -> OtherResult<Vec<Signature>>

This gets a list of all signers so far on a multisigtx.

pub fn pause(&mut self) -> OtherResult<()>

Function pauses contract.

Any signatory may call.

pub fn unpause(&mut self, function: String) -> OtherResult<()>

Function unpauses contract.

pub fn reward_interlocker( &mut self, reward: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, interlocker: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> OtherResult<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>

Reward the interlocker for browsing, etc.

This is a manual rewarding function, to override the socket formalism.

pub fn rewarded_interlocker_total( &self, interlocker: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> OtherResult<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>

Get amount rewarded to interlocker to date.

pub fn rewarded_total( &self ) -> <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance

Get total amount rewarded to date.

pub fn transfer_ownership( &mut self, newowner: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, function: String ) -> Result<(), OtherError>

pub fn update_contract( &mut self, code_hash: [u8; 32], function: String ) -> OtherResult<()>

Modifies the code which is used to execute calls to this contract address.

This upgrades the token contract logic while using old state.

pub fn create_port( &mut self, codehash: <<ILOCKmvp as ContractEnv>::Env as Environment>::Hash, tax: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, cap: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, locked: bool, number: u16, owner: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, overwrite: bool, function: String ) -> OtherResult<()>

Create a new port that application contract can register with.

Each port tracks amount rewarded, tax collected, if it is locked or not, owner.

A locked port may only be registered by the Interlock Network foundation.

pub fn create_socket( &mut self, operator: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, portnumber: u16 ) -> OtherResult<()>

Rewards/staking/application contracts register with this token contract here.

Contract must first register with token contract as port to allow connection via socket (ie, a port must first exist before a socket may form)..

pub fn call_socket( &mut self, address: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, amount: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, _data: Vec<u8> ) -> OtherResult<()>

Check for socket and apply custom logic after being called from application contract.

Application contract calls its socket to trigger internal logic defined by port.

Default parameters are address and amount or value.

Additional parameters may be encoded as _data: Vec.

pub fn socket( &self, application: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> Socket

Get socket info.

pub fn port(&self, portnumber: u16) -> Port

Get port info.

pub fn emit_event<EE: EmitEvent<Self>>(emitter: EE, event: Event)

Function for internal _emit_event implementations.

pub fn check_multisig(&mut self, function: String) -> OtherResult<()>

Helper function for checking signature count

pub fn tax_port_transfer( &mut self, socket: Socket, port: Port, amount: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance ) -> OtherResult<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>

Tax and reward transfer between socket calling address and socket operator.