Struct ilockmvp::ilockmvp::Socket

pub struct Socket {
    operator: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId,
    portnumber: u16,
}

Information pertaining to socket definition in application/socket/port contract connectivity formalism.

Fields
operator: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId

Who operates (owns usually) a specific instance of a connecting application contract?

Using the restaurant franchise metaphor again, the operator may have several different instances of the port’s application contract.

Each instance of the application contract has its own address, but each restaurant has the same operator.

The socket (operator:franchise or operator:port#) is like the single business franchise agreement between the restaurant operator and the franchise owner.

There is only one agreement between the franchise and the restaurant operator, regardless of how many restaurants the operator has.

portnumber: u16

What port is this operator connected to?

Using the restaurant franchise metaphor again, the port is like the franchise itself.

The port number is what identifies a particular franchise and its protocols, procedures, metadata, and ultimately business model and standards for any franchisees.