Struct ilockmvp::ilockmvp::VestData

pub struct VestData {
    pub stakeholder: <Mapping<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, StakeholderData> as AutoStorableHint<ManualKey<{ _ }, ManualKey<VEST_KEY>>>>::Type,
    pub monthspassed: <u16 as AutoStorableHint<ManualKey<{ _ }, ManualKey<VEST_KEY>>>>::Type,
    pub nextpayout: <<<ILOCKmvp as ContractEnv>::Env as Environment>::Timestamp as AutoStorableHint<ManualKey<{ _ }, ManualKey<VEST_KEY>>>>::Type,
    pub _reserved: <Option<()> as AutoStorableHint<ManualKey<{ _ }, ManualKey<VEST_KEY>>>>::Type,
}

Fields
stakeholder: <Mapping<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, StakeholderData> as AutoStorableHint<ManualKey<{ _ }, ManualKey<VEST_KEY>>>>::Type

Contains information about stakeholders and the vesting status.

See detailed struct below.

stakeholder: stakeholder account address -> info about stakeholder

monthspassed: <u16 as AutoStorableHint<ManualKey<{ _ }, ManualKey<VEST_KEY>>>>::Type

Counter responsible for keeping track of how many months have passed along the vesting schedule.

Used in part to calculate and compare token amount paid out vs token amount owed.

nextpayout: <<<ILOCKmvp as ContractEnv>::Env as Environment>::Timestamp as AutoStorableHint<ManualKey<{ _ }, ManualKey<VEST_KEY>>>>::Type

Stores the date timestamp one month ahead of the last increment of monthspassed

_reserved: <Option<()> as AutoStorableHint<ManualKey<{ _ }, ManualKey<VEST_KEY>>>>::Type

Expand storage related to the vesting functionality.