Module ilockmvp::ilockmvp

Structs

AppData

Approval

specify approve event

CallBuilder

The ink! smart contract’s call builder.

Check 🔒

ILOCKmvp

ILOCKmvp struct contains overall storage data for contract

ILOCKmvpRef

ILOCKmvp struct contains overall storage data for contract

PoolData

Port

Information pertaining to port definition in application/socket/port contract connectivity formalism.

Reward

specify reward event

RewardData

Socket

Information pertaining to socket definition in application/socket/port contract connectivity formalism.

StakeholderData

StakeholderData struct contains all pertinent information for each stakeholder (Besides balance and allowance mappings).

This is primarily for managing and implementing the vesting schedule.

TokenPools

Transfer

specify transfer event

VestData

Enums

OtherError

Other contract error types

__ink_ConstructorDecoder

__ink_EventBase

__ink_MessageDecoder

__ink_UndefinedAmountOfTopics

Constants

ADVISORS

APP_KEY

This is upgradable storage for the application connection feature of this PSP22 contract (ie, the application/socket/port contract connectivity formalism).

ARR_LEN 🔒

ARR_LEN 🔒

ARR_LEN 🔒

ARR_LEN 🔒

CONCATP_NHPMWYD3NJA 🔒

CONCATP_NHPMWYD3NJA 🔒

CONCATP_NHPMWYD3NJA 🔒

CONCATP_NHPMWYD3NJA 🔒

CONCAT_ARR 🔒

CONCAT_ARR 🔒

CONCAT_ARR 🔒

CONCAT_ARR 🔒

CONCAT_STR 🔒

CONCAT_STR 🔒

CONCAT_STR 🔒

CONCAT_STR 🔒

CONSTRUCTOR_0 🔒

DECIMALS_POWER10

EARLY_BACKERS

FOUNDATION

ID_LENGTH

magic numbers

MESSAGE_0 🔒

MESSAGE_1 🔒

MESSAGE_2 🔒

MESSAGE_3 🔒

MESSAGE_4 🔒

MESSAGE_5 🔒

MESSAGE_6 🔒

MESSAGE_7 🔒

MESSAGE_8 🔒

MESSAGE_9 🔒

MESSAGE_10 🔒

MESSAGE_11 🔒

MESSAGE_12 🔒

MESSAGE_13 🔒

MESSAGE_14 🔒

MESSAGE_15 🔒

MESSAGE_16 🔒

MESSAGE_17 🔒

MESSAGE_18 🔒

MESSAGE_19 🔒

MESSAGE_20 🔒

MESSAGE_21 🔒

MESSAGE_22 🔒

MESSAGE_23 🔒

MESSAGE_24 🔒

MESSAGE_25 🔒

MESSAGE_26 🔒

MESSAGE_27 🔒

MESSAGE_28 🔒

MESSAGE_29 🔒

MESSAGE_30 🔒

MESSAGE_31 🔒

MESSAGE_32 🔒

MESSAGE_33 🔒

MESSAGE_34 🔒

MESSAGE_35 🔒

MESSAGE_36 🔒

MESSAGE_37 🔒

ONE_MONTH

OUTLIER_VENTURES

PARTNERS

POOLS

pool data

POOL_COUNT

POOL_KEY

This is upgradable storage for the token pool management and accounting feature of this PSP22 contract.

PRESALE_1

PRESALE_2

PRESALE_3

PUBLIC_SALE

REWARDS

REWARD_KEY

This is upgradable storage for the token rewarding feature of this PSP22 contract.

SUPPLY_CAP

TEAM_FOUNDERS

TOKEN_CAP

token data

TOKEN_DECIMALS

TOKEN_NAME

TOKEN_SYMBOL

VEST_KEY

This is upgradable storage for the application connection feature of this PSP22 contract (ie, the application/socket/port contract connectivity formalism).

WHITELIST

Functions

__ink_generate_metadata

call 🔒

deploy 🔒

push_contract 🔒

Type Definitions

AccountId 🔒

Balance 🔒

BlockNumber 🔒

Environment 🔒

Event

Hash 🔒

OtherResult

PSP22Result

Timestamp 🔒