Struct ilockmvp::ilockmvp::CallBuilder

#[repr(transparent)]
pub struct CallBuilder {
    account_id: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId,
}

The ink! smart contract’s call builder.

Implements the underlying on-chain calling of the ink! smart contract messages and trait implementations in a type safe way.

Fields
account_id: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId
Implementations

impl CallBuilder

pub fn order_multisigtx( &mut self, __ink_binding_0: String ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, EmptyArgumentList>>>, Set<ReturnType<OtherResult<()>>>>

Function to order multisigtx transaction.

pub fn sign_multisigtx( &mut self, __ink_binding_0: String ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, EmptyArgumentList>>>, Set<ReturnType<OtherResult<()>>>>

A multisigtx signer calls this to sign.

pub fn add_signatory( &mut self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, __ink_binding_1: String ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>>, Set<ReturnType<OtherResult<()>>>>

This adds a signatory from the list of permitted signatories.

pub fn remove_signatory( &mut self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, __ink_binding_1: String ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>>, Set<ReturnType<OtherResult<()>>>>

This removes a signatory from the list of permitted signatories.

pub fn change_threshold( &mut self, __ink_binding_0: u16, __ink_binding_1: String ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, ArgumentList<Argument<u16>, EmptyArgumentList>>>>, Set<ReturnType<OtherResult<()>>>>

This changes signer threshold for approving multisigtx.

pub fn change_multisigtxtimelimit( &mut self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::Timestamp, __ink_binding_1: String ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::Timestamp>, EmptyArgumentList>>>>, Set<ReturnType<OtherResult<()>>>>

This modifies timelimit for a multisig transaction.

pub fn threshold( &self ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<u16>>>

This gets the current signature threshold for multisigtx.

pub fn multisigtimelimit( &self ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<<<ILOCKmvp as ContractEnv>::Env as Environment>::Timestamp>>>

This gets the current timelimit for signatories to sign multisigtx.

pub fn signatories( &mut self ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<OtherResult<Vec<AccountID>>>>>

This gets a list of current accounts permitted to sign multisigtx.

pub fn signatory_count( &self ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<u8>>>

This gets number of signatories permitted to sign multisigtx.

pub fn signature_count( &self ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<u8>>>

This gets current number of signatures for multisigtx.

pub fn check_signatures( &mut self ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<OtherResult<Vec<Signature>>>>>

This gets a list of all signers so far on a multisigtx.

pub fn pause( &mut self ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<OtherResult<()>>>>

Function pauses contract.

Any signatory may call.

pub fn unpause( &mut self, __ink_binding_0: String ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, EmptyArgumentList>>>, Set<ReturnType<OtherResult<()>>>>

Function unpauses contract.

pub fn reward_interlocker( &mut self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, __ink_binding_1: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>, EmptyArgumentList>>>>, Set<ReturnType<OtherResult<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>>>>

Reward the interlocker for browsing, etc.

This is a manual rewarding function, to override the socket formalism.

pub fn rewarded_interlocker_total( &self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>, Set<ReturnType<OtherResult<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>>>>

Get amount rewarded to interlocker to date.

pub fn rewarded_total( &self ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>>>

Get total amount rewarded to date.

pub fn transfer_ownership( &mut self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, __ink_binding_1: String ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>>, Set<ReturnType<Result<(), OtherError>>>>

pub fn update_contract( &mut self, __ink_binding_0: [u8; 32], __ink_binding_1: String ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, ArgumentList<Argument<[u8; 32]>, EmptyArgumentList>>>>, Set<ReturnType<OtherResult<()>>>>

Modifies the code which is used to execute calls to this contract address.

This upgrades the token contract logic while using old state.

pub fn create_port( &mut self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::Hash, __ink_binding_1: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, __ink_binding_2: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, __ink_binding_3: bool, __ink_binding_4: u16, __ink_binding_5: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, __ink_binding_6: bool, __ink_binding_7: String ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, ArgumentList<Argument<bool>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, ArgumentList<Argument<u16>, ArgumentList<Argument<bool>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::Hash>, EmptyArgumentList>>>>>>>>>>, Set<ReturnType<OtherResult<()>>>>

Create a new port that application contract can register with.

Each port tracks amount rewarded, tax collected, if it is locked or not, owner.

A locked port may only be registered by the Interlock Network foundation.

pub fn create_socket( &mut self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, __ink_binding_1: u16 ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<u16>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>>, Set<ReturnType<OtherResult<()>>>>

Rewards/staking/application contracts register with this token contract here.

Contract must first register with token contract as port to allow connection via socket (ie, a port must first exist before a socket may form)..

pub fn call_socket( &mut self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, __ink_binding_1: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, __ink_binding_2: Vec<u8> ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<Vec<u8>>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>, ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>>>, Set<ReturnType<OtherResult<()>>>>

Check for socket and apply custom logic after being called from application contract.

Application contract calls its socket to trigger internal logic defined by port.

Default parameters are address and amount or value.

Additional parameters may be encoded as _data: Vec.

pub fn socket( &self, __ink_binding_0: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<<<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>, Set<ReturnType<Socket>>>

Get socket info.

pub fn port( &self, __ink_binding_0: u16 ) -> CallBuilder<<ILOCKmvp as ContractEnv>::Env, Set<Call<<ILOCKmvp as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<u16>, EmptyArgumentList>>>, Set<ReturnType<Port>>>

Get port info.